update Login.js file and login action creator; add separate businessActions file
- [Login] move businessAccount boolean in credentials object
- [Actions] update login action creator to pass new object with only email and password
- [Actions] add Promise.reject to all action creators (this properly handles errors before reaching a chained .then() from another file)
- [Actions] create separate businessActions file
- [BusinessDash] update mapStateToProps
- [BusinessDash] add minor styling
