+++ replate/src/components/Business/BusinessDash.css
@@ -11,6 +11,25 @@
     margin-bottom: 5%;
 }
 
+.add-btn-wrapper {
+    display: flex;
+}
+
 .donation-request-wrapper {
     display: flex;
+}
+
+ul {
+    list-style: none;
+}
+
+li {
+
+    padding-bottom: 8px;
+}
+
+.location-name {
+    border-bottom: 1px solid;
+    margin-bottom: 10px;
+    padding-bottom: 10px;
 }
